Feat: 레포 뷰 조회 기능 추가
레포 뷰 조회 기능 추가

+++ backend/src/main/java/com/dev101/coa/domain/code/dto/CodeCntDto.java
@@ -7,5 +7,5 @@ import lombok.Getter;
 @Builder
 public class CodeCntDto {
     private String codeName;
-    private Long lineCnt;
+    private Integer lineCnt;
 }
